로그아웃하는 로직 모달 -> 헤더
