--- src/Hooks/useHistoryDetail.js
@@ -1,6 +1,0 @@
-import { useEffect, useState } from "react";
-
-const useHistoryDetail = () => {
-};
-
-export default useHistoryDetail;
