+++ src/pages/HistoryDetail/index.jsx
@@ -1,14 +1,12 @@
 import React from "react";
 import * as C from "../../components";
-import useHistoryDetail from "../../Hooks/useHistoryDetail";
 
 const HistoryDetail = () => {
-  const {title} = useHistoryDetail();
   return (
     <>
       <C.RecentModified />
       <C.Header />
-      <C.PageContainer title={title} sort="역사">
+      <C.PageContainer title={"타이틀이요"} sort="역사">
         <C.Explanation>
           <C.HistoryDetail />
         </C.Explanation>
